+++ src/app/edit-project-form/edit-project-form.component.ts
@@ -1,24 +1,33 @@
 import { Component, Input, ChangeDetectionStrategy } from '@angular/core';
 import { Router } from '@angular/router';
 
+
 import { Project } from '../project.js';
 import { ProjectService } from '../project.service.js';
+import { ConfirmationService } from 'primeng/api';
 
 @Component({
   selector: 'app-edit-project-form',
   templateUrl: './edit-project-form.component.html',
   styleUrl: './edit-project-form.component.scss',
-  changeDetection: ChangeDetectionStrategy.OnPush
+  changeDetection: ChangeDetectionStrategy.OnPush,
 })
 export class EditProjectFormComponent {
   @Input() project?: Project;
 
   submitted = false;
 
+  projectHasTickets = true;
+
   constructor(
     private projectService: ProjectService,
     private router: Router,
+    private confirmationService: ConfirmationService,
   ) { }
+  
+  ngOnInit(): void {
+    this.checkTickets();
+  }
 
   onSubmit() {
     this.submitted = true;
@@ -33,11 +42,30 @@ export class EditProjectFormComponent {
     }
   }
 
-  deleteProject(): void {
+  checkTickets(): void {
     if (this.project) {
-      this.projectService.deleteProject(this.project.id)
-      .subscribe(_ => this.router.navigate(['/projects']));
+      if (this.project.tickets.length === 0) {
+        this.projectHasTickets = false;
+      }
     }
   }
 
+  // deleteProject(): void {
+  //   if (this.project) {
+  //     this.projectService.deleteProject(this.project.id)
+  //     .subscribe(_ => this.router.navigate(['/projects']));
+  //   }
+  // }
+
+  deleteProject() {
+    this.confirmationService.confirm({
+        message: '¿Confirma que desea eliminar el proyecto?',
+        accept: () => {
+          if (this.project) {
+            this.projectService.deleteProject(this.project.id)
+            .subscribe(_ => this.router.navigate(['/projects']));
+          }
+        }
+    });
+  }
 }
